pull from upstream

+++ public/app/features/folders/FolderSettingsPage.tsx
@@ -1,13 +1,16 @@
 import React, { PureComponent } from 'react';
 import { hot } from 'react-hot-loader';
 import { connect } from 'react-redux';
-import { Input, NavModel } from '@grafana/ui';
+import { NavModel } from '@grafana/data';
+import { LegacyForms } from '@grafana/ui';
+const { Input } = LegacyForms;
 import Page from 'app/core/components/Page/Page';
 import appEvents from 'app/core/app_events';
 import { getNavModel } from 'app/core/selectors/navModel';
-import { StoreState, FolderState } from 'app/types';
-import { getFolderByUid, setFolderTitle, saveFolder, deleteFolder } from './state/actions';
+import { CoreEvents, FolderState, StoreState } from 'app/types';
+import { deleteFolder, getFolderByUid, saveFolder } from './state/actions';
 import { getLoadingNav } from './state/navModel';
+import { setFolderTitle } from './state/reducers';
 
 export interface Props {
   navModel: NavModel;
@@ -51,10 +54,10 @@ export class FolderSettingsPage extends PureComponent<Props, State> {
     evt.stopPropagation();
     evt.preventDefault();
 
-    appEvents.emit('confirm-modal', {
+    appEvents.emit(CoreEvents.showConfirmModal, {
       title: 'Delete',
       text: `Do you want to delete this folder and all its dashboards?`,
-      icon: 'fa-trash',
+      icon: 'trash-alt',
       yesText: 'Delete',
       onConfirm: () => {
         this.props.deleteFolder(this.props.folder.uid);
@@ -114,9 +117,4 @@ const mapDispatchToProps = {
   deleteFolder,
 };
 
-export default hot(module)(
-  connect(
-    mapStateToProps,
-    mapDispatchToProps
-  )(FolderSettingsPage)
-);
+export default hot(module)(connect(mapStateToProps, mapDispatchToProps)(FolderSettingsPage));
